All frontend default data is now a json
