--- frontend/src/utils/default-data/nonProfitDefaultData.js
@@ -1,13 +1,0 @@
-// All nonprofit default data
-export const nonprofitDefaultOption = {
-  id: "unselected",
-  text: "Not Selected",
-};
-
-// Default data for the Stats of a Non Profit
-export const statsDefault = {
-  servicesNumber: { name: "Services #", key: "null" },
-  servicesOffered: { name: "Services offered", key: "null" },
-  popularZipCode: { name: "Popular Zip Codes", key: "null" },
-  servicesOffered2: { name: "Cost", key: "null" },
-};
