+++ frontend/src/utils/default-data/nonprofitDefaultOption.json
@@ -1,0 +1,4 @@
+{
+  "id": "unselected",
+  "text": "Not Selected"
+}
