+++ frontend/src/utils/default-data/serviceNameInputDefault.json
@@ -1,0 +1,4 @@
+{
+  "id": "unselected",
+  "text": "Select a Service"
+}
