+++ frontend/src/utils/default-data/statsDefault.json
@@ -1,0 +1,6 @@
+{
+  "servicesNumber": { "name": "Services #", "key": "null" },
+  "servicesOffered": { "name": "Services offered", "key": "null" },
+  "popularZipCode": { "name": "Popular Zip Codes", "key": "null" },
+  "servicesOffered2": { "name": "Cost", "key": "null" }
+}
