+++ frontend/src/utils/serviceIconUtils.js
@@ -1,0 +1,37 @@
+import {
+  MdOutlinePerson,
+  MdOutlineLocationOn,
+  MdNumbers,
+  MdOutlineTextSnippet,
+  MdEmail,
+  MdLocalPhone,
+  MdLanguage,
+  MdHourglassBottom,
+  MdImage,
+  MdOutlineSpatialAudioOff,
+  MdDateRange,
+  MdFilterList,
+  MdOutlineStopCircle,
+  MdOutlineSkipNext,
+  MdAirplanemodeActive,
+  MdOutlinePersonAddAlt,
+} from "react-icons/md";
+
+export const serviceSearchIconMap = {
+  MdOutlineLocationOn: MdOutlineLocationOn,
+  MdOutlineStopCircle: MdOutlineStopCircle,
+  MdAirplanemodeActive: MdAirplanemodeActive,
+  MdOutlineSpatialAudioOff: MdOutlineSpatialAudioOff,
+  MdDateRange: MdDateRange,
+  MdOutlinePersonAddAlt: MdOutlinePersonAddAlt,
+  MdOutlinePerson: MdOutlinePerson,
+  MdNumbers: MdNumbers,
+  MdOutlineTextSnippet: MdOutlineTextSnippet,
+  MdEmail: MdEmail,
+  MdLocalPhone: MdLocalPhone,
+  MdLanguage: MdLanguage,
+  MdHourglassBottom: MdHourglassBottom,
+  MdImage: MdImage,
+  MdFilterList: MdFilterList,
+  MdOutlineSkipNext: MdOutlineSkipNext,
+};
